requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
ollama>=0.3.0
speechmatics-python>=1.6.0
python-dotenv>=1.0.0